fix: set buffer options before opening window
I was thinking that maybe the ordering was the reason why LSP clients
(mainly otter.nvim) was insisting on attaching to the buffer even though
the buffer options should've prevented it in my personal config. It
ended up not being the case, but it still might be a good idea in case
it affects something else down the line.